Moving and creating result component
